safari-600.7-branch/LayoutTests/fast/canvas/canvas-webkitLineDash-expected.txt

Basic test for webkitLineDash and webkitLineDashOffset

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.webkitLineDashOffset is 0
PASS lineDash[0] is 15
PASS lineDash[1] is 10
PASS ctx.webkitLineDashOffset is 5
PASS ctx.webkitLineDashOffset is 5
PASS successfullyParsed is true

TEST COMPLETE
